pyproject.toml
[tool.poetry]
name = "5-cw"
version = "0.1.0"
description = ""
authors = ["Анатолий Шаповалов <anatoly.shapovalov@gmail.com>"]
readme = "README.md"
packages = [{include = "5_cw"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.2"
psycopg2-binary = "^2.9.6"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
